pyproject.toml
[project]
name = "ai-coding-agent"
version = "0.1.0"
description = "Multi-provider AI coding agent supporting Gemini, Codex, and Claude"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "AI Coding Agent Team"}
]
keywords = ["ai", "coding", "agent", "gemini", "codex", "claude", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "pyyaml>=6.0",
    "httpx>=0.25",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "litellm>=1.0",
    "google-generativeai>=0.3",
    "anthropic>=0.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
    "mypy>=1.0",
]
browser = [
    "playwright>=1.40",
    "patchright>=1.0",
    "websocket-client>=1.6",
]

[project.scripts]
aca = "ai_coding_agent.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_coding_agent"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
strict = true
